Extract mixed-language context tests
Extract `DocumentationContext` tests relating to multi-language support
to a separate file.
